requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
greenlet==3.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.5
